Dev : Refactor Changed Validator. Moved the factory part on a new ValidatorFactory
