+++ src/main/java/com/tech/configurations/tools/Validator.java
@@ -5,29 +5,19 @@
  */
 package com.tech.configurations.tools;
 
-import com.tech.configurations.tools.customvalidators.interfaces.CustomValidator;
-import com.tech.models.dtos.superclass.BaseDTO;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 /**
  *
  * @author KuroiTenshi
  */
 public class Validator {
-    private static final HashMap<Integer,CustomValidator> AVAILABLE_VALIDATORS = new HashMap();
     private static final List<String> AVAILABLE_PRIVILEGES = new ArrayList();
     private static final List<String> AVAILABLE_METHODS = new ArrayList();
-    private static ResponseEntity response = new ResponseEntity(HttpStatus.NOT_MODIFIED);
-    public static ResponseEntity getResponse() {
-        return response;
-    }
-    
+
     /**
      * Tests a name if the formations / data are acceptable for the program
      * General nameValidator. for specific validations write your own :)
@@ -98,10 +88,10 @@ public class Validator {
     }
    
     /**
-    * 
-    * @param accessMethod
-    * @return returns true if AccessMethod is either blacklist or whitelist
-    */
+     * 
+     * @param accessMethod
+     * @return returns true if AccessMethod is either blacklist or whitelist
+     */
     public static boolean accessMethodValidator(String accessMethod){
         if(accessMethod.isEmpty()){
             return false;
@@ -128,6 +118,7 @@ public class Validator {
         }
         return false;
     }
+    
     /**
      * General idValidator. for specific validations write your own :)
      * @param id
@@ -142,24 +133,6 @@ public class Validator {
             return false;
         }
         
-        return true;
-    }
-    
-    public static Pair validateDTO (BaseDTO DTO) {
-        int code = DTO.getDTOName().toLowerCase().hashCode();
-        if (AVAILABLE_VALIDATORS.containsKey(code)) {
-            response = new ResponseEntity(Responses.VALIDATOR_NOT_FOUND.getData(),HttpStatus.NOT_FOUND);
-            return new Pair(false, response);
-        }
-        boolean booleanResponse = AVAILABLE_VALIDATORS.get(code).validate(DTO);
-        return new Pair(booleanResponse, AVAILABLE_VALIDATORS.get(code).response());
-    }
-    
-    public static boolean registerDTOValidator(CustomValidator DTOvalidator) {
-        if (AVAILABLE_VALIDATORS.containsKey(DTOvalidator.hashCode())) {
-            return false;
-        }
-        AVAILABLE_VALIDATORS.put(DTOvalidator.hashCode(),DTOvalidator);
         return true;
     }
